// Electron/src/swagger/api/models/Genre.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Song genres
 */
export enum Genre {
    POP = 'Pop',
    ROCK = 'Rock',
    HIP_HOP = 'Hip-hop',
    R_B_RITMO_Y_BLUES_ = 'R&B (Ritmo y Blues)',
    JAZZ = 'Jazz',
    BLUES = 'Blues',
    REGGAE = 'Reggae',
    COUNTRY = 'Country',
    FOLK = 'Folk',
    CL_SICA = 'Clásica',
    ELECTR_NICA = 'Electrónica',
    DANCE = 'Dance',
    METAL = 'Metal',
    PUNK = 'Punk',
    FUNK = 'Funk',
    SOUL = 'Soul',
    GOSPEL = 'Gospel',
    LATINA = 'Latina',
    M_SICA_DEL_MUNDO = 'Música del mundo',
    EXPERIMENTAL = 'Experimental',
    AMBIENTAL = 'Ambiental',
    FUSI_N = 'Fusión',
    INSTRUMENTAL = 'Instrumental',
    ALTERNATIVA = 'Alternativa',
    INDIE = 'Indie',
    RAP = 'Rap',
    SKA = 'Ska',
    GRUNGE = 'Grunge',
    TRAP = 'Trap',
    REGGAETON = 'Reggaeton',
}